requirements.txt
fastapi
uvicorn
SQLAlchemy

psycopg2
asyncpg
alembic
pydantic-settings